Setting Transition Values
I have implemented what should take the value from the radio button
selected and play that transition for the whole show once the play
button is clicked. It does not seem to be storing the right values so if
someone could run it and give me some feedback that would be great!