Made changes to GUI
Test it out and let me know what you think.(might not look the same on
every computer). Changed the player background to gray so that the fade
using black and white are more easily visible. Slider is now at the
bottom and undecorated so it blends in better. Also we can start adding
jpg images to the pictures folder whenever.